src/app/scripts/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
